Funciones para crear, editar, eliminar, y listar menu modularizadas

+++ client/src/utils/useMenu.ts
@@ -1,25 +1,22 @@
-import { useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import Swal from "sweetalert2";
 import { adminService } from "@/services/adminService";
 import { MenuItem, MenuCategory } from "@/types/MenuItem";
+import { UsersContext } from "@/context/UsersContext";
 
 export function useMenu() {
-  const [menuItems, setMenuItems] = useState<MenuItem[]>([]);
-  const [categories, setCategories] = useState<MenuCategory[]>([]);
-  const [loading, setLoading] = useState(true);
+  // const [menuItems, setMenuItems] = useState<MenuItem[]>([]);
+  const [platesList, setPlatesList] = useState<MenuCategory[]>([]);
+  const { isLogged } = useContext(UsersContext);
 
   const fetchMenu = async () => {
     try {
-      setLoading(true);
-      const { categories, menuItems } =
-        await adminService.getAllPlates();
+      const { categories } = await adminService.getAllPlates();
 
-      setCategories(categories);
-      setMenuItems(menuItems);
-    } catch {
+      setPlatesList(categories);
+    } catch (error) {
+      console.error(error);
       Swal.fire("Error", "No se pudo cargar el menú", "error");
-    } finally {
-      setLoading(false);
     }
   };
 
@@ -27,9 +24,12 @@ export function useMenu() {
     fetchMenu();
   }, []);
 
-  const createItem = async (item: Omit<MenuItem, "id">) => {
+  const createItem = async (item: any) => {
     try {
-      await adminService.createNewPlate(item);
+      await adminService.createNewPlate({
+        ...item,
+        restaurant_id: isLogged?.restaurant_id,
+      });
       await fetchMenu();
       Swal.fire("Éxito", "Platillo creado", "success");
     } catch {
@@ -37,18 +37,41 @@ export function useMenu() {
     }
   };
 
-  const updateItem = async (item: MenuItem) => {
+  const updateItem = async (item: any) => {
     try {
-      await adminService.updatePlateInfo(item.id, item);
+      const restaurant_id =
+        "11111111-1111-1111-1111-111111111111"; /*isLogged?.restaurant_id,*/
+
+      if (!restaurant_id) {
+        throw new Error("No hay restaurant_id");
+      }
+
+      await adminService.updatePlateInfo(item.id, item, restaurant_id);
+
       await fetchMenu();
+
       Swal.fire("Actualizado", "Cambios guardados", "success");
-    } catch {
+    } catch (error) {
+      console.error(error);
       Swal.fire("Error", "No se pudo actualizar", "error");
     }
   };
 
   const deleteItem = async (id: string) => {
     try {
+      const result = await Swal.fire({
+        title: "¿Eliminar producto?",
+        text: "El producto dejará de estar disponible en el menú.",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#d33",
+        cancelButtonColor: "#6b7280",
+        confirmButtonText: "Sí, eliminar",
+        cancelButtonText: "Cancelar",
+      });
+
+      if (!result.isConfirmed) return;
+
       await adminService.deletePlate(id);
       await fetchMenu();
       Swal.fire("Eliminado", "Platillo eliminado", "success");
@@ -57,26 +80,25 @@ export function useMenu() {
     }
   };
 
-  const changeStatus = async (
-    id: string,
-    status: "disponible" | "agotado" | "inactivo"
-  ) => {
-    try {
-      await adminService.updatePlateStatusFlexible(id, status);
-      await fetchMenu();
-    } catch {
-      Swal.fire("Error", "No se pudo actualizar estado", "error");
-    }
-  };
+  // const changeStatus = async (
+  //   id: string,
+  //   status: "disponible" | "agotado" | "inactivo"
+  // ) => {
+  //   try {
+  //     await adminService.updatePlateStatusFlexible(id, status);
+  //     await fetchMenu();
+  //   } catch {
+  //     Swal.fire("Error", "No se pudo actualizar estado", "error");
+  //   }
+  // };
 
   return {
-    menuItems,
-    categories,
-    loading,
+    // menuItems,
+    platesList,
     createItem,
     updateItem,
     deleteItem,
-    changeStatus,
-    refetch: fetchMenu,
+    // changeStatus,
+    // refetch: fetchMenu,
   };
 }
